chore(connect-explorer-theme): another step removing circulars
