+++ packages/connect-explorer-theme/src/components/banner.tsx
@@ -3,7 +3,7 @@ import type { ReactElement } from 'react';
 import cn from 'clsx';
 import { XIcon } from 'nextra/icons';
 
-import { useConfig } from '../contexts/config';
+import { useConfig } from '../contexts/useConfig';
 import { renderComponent } from '../utils';
 
 export function Banner(): ReactElement | null {
