+++ packages/connect-explorer-theme/src/components/nav-links.tsx
@@ -4,7 +4,7 @@ import cn from 'clsx';
 import { ArrowRightIcon } from 'nextra/icons';
 import type { Item } from 'nextra/normalize-pages';
 
-import { useConfig } from '../contexts/config';
+import { useConfig } from '../contexts/useConfig';
 import type { DocsThemeConfig } from '../index';
 import { Anchor } from './anchor';
