+++ packages/connect-explorer-theme/src/components/navbar.tsx
@@ -11,8 +11,8 @@ import { useElevation } from '@trezor/components';
 import { TrezorLogo } from '@trezor/product-components';
 import { Elevation, borders, mapElevationToBackground, spacingsPx } from '@trezor/theme';
 
-import { useConfig } from '../contexts/config';
 import { useMenu } from '../contexts/menu';
+import { useConfig } from '../contexts/useConfig';
 import { renderComponent } from '../utils';
 import { Anchor } from './anchor';
