+++ packages/connect-explorer-theme/src/components/theme-switch.tsx
@@ -8,7 +8,7 @@ import { z } from 'zod';
 
 import { Select } from '@trezor/components';
 
-import { useConfig } from '../contexts/config';
+import { useConfig } from '../contexts/useConfig';
 
 type ThemeSwitchProps = {
     lite?: boolean;
