+++ packages/connect-explorer-theme/src/contexts/useConfig.tsx
@@ -1,0 +1,11 @@
+import { createContext, useContext } from 'react';
+
+import type { FrontMatter, PageOpts } from 'nextra';
+
+import type { DocsThemeConfig } from '../schema';
+export type Config<FrontMatterType = FrontMatter> = DocsThemeConfig &
+    Pick<PageOpts<FrontMatterType>, 'flexsearch' | 'newNextLinkBehavior' | 'title' | 'frontMatter'>;
+export const ConfigContext = createContext<Config>({} as Config);
+export function useConfig<FrontMatterType = FrontMatter>() {
+    return useContext<Config<FrontMatterType>>(ConfigContext);
+}
