+++ packages/connect-explorer-theme/src/index.tsx
@@ -15,10 +15,12 @@ import { ElevationContext } from '@trezor/components';
 import { Elevation, mapElevationToBackground } from '@trezor/theme';
 
 import { Banner, Breadcrumb, Head, NavLinks, Navbar, Sidebar, SkipNavContent } from './components';
-import { DEFAULT_LOCALE, PartialDocsThemeConfig } from './constants';
+import { DEFAULT_LOCALE } from './constants';
 import { ActiveAnchorProvider } from './contexts/active-anchor';
-import { ConfigProvider, useConfig } from './contexts/config';
+import { ConfigProvider } from './contexts/config';
+import { useConfig } from './contexts/useConfig';
 import { getComponents } from './mdx-components';
+import { PartialDocsThemeConfig } from './schema';
 import { patchedNormalizePages, renderComponent } from './utils';
 
 const GlobalStyle = createGlobalStyle<{ $elevation: Elevation }>`
