+++ packages/connect-explorer-theme/src/types.ts
@@ -2,7 +2,7 @@ import type { ReactNode } from 'react';
 
 import type { PageOpts } from 'nextra';
 
-import type { DocsThemeConfig } from './constants';
+import type { DocsThemeConfig } from './schema';
 
 export type Context = {
     pageOpts: PageOpts;
