+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -30,24 +30,15 @@ packages/components/src/components/form/Select/Select.tsx > packages/components/
 packages/components/src/index.ts > packages/components/src/components/form/SelectBar/SelectBar.tsx
 packages/components/src/index.ts > packages/components/src/support/Story.tsx
 packages/components/src/utils/frameProps.tsx > packages/components/src/components/Flex/Flex.tsx
-packages/connect-explorer-theme/src/components/anchor.tsx > packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx
-packages/connect-explorer-theme/src/components/anchor.tsx > packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/flexsearch.tsx > packages/connect-explorer-theme/src/components/search.tsx
-packages/connect-explorer-theme/src/components/anchor.tsx > packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/navbar.tsx
-packages/connect-explorer-theme/src/components/anchor.tsx > packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/toc.tsx
+packages/connect-explorer-theme/src/components/anchor.tsx > packages/connect-explorer-theme/src/contexts/useConfig.tsx > packages/connect-explorer-theme/src/schema.ts > packages/connect-explorer-theme/src/components/navbar.tsx
+packages/connect-explorer-theme/src/components/anchor.tsx > packages/connect-explorer-theme/src/contexts/useConfig.tsx > packages/connect-explorer-theme/src/schema.ts > packages/connect-explorer-theme/src/components/toc.tsx
 packages/connect-explorer-theme/src/components/index.ts > packages/connect-explorer-theme/src/components/nav-links.tsx > packages/connect-explorer-theme/src/index.tsx
 packages/connect-explorer-theme/src/components/index.ts > packages/connect-explorer-theme/src/components/nav-links.tsx > packages/connect-explorer-theme/src/index.tsx > packages/connect-explorer-theme/src/mdx-components.tsx
 packages/connect-explorer-theme/src/components/sidebar.tsx > packages/connect-explorer-theme/src/components/menu.tsx
-packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/flexsearch.tsx
-packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/flexsearch.tsx > packages/connect-explorer-theme/src/components/search.tsx > packages/connect-explorer-theme/src/types.ts
-packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/toc.tsx > packages/connect-explorer-theme/src/contexts/active-anchor.tsx
-packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx
-packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/flexsearch.tsx > packages/connect-explorer-theme/src/components/search.tsx
-packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/flexsearch.tsx > packages/connect-explorer-theme/src/components/search.tsx > packages/connect-explorer-theme/src/utils/index.ts > packages/connect-explorer-theme/src/utils/use-git-edit-url.ts
-packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/footer.tsx
-packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/footer.tsx > packages/connect-explorer-theme/src/components/locale-switch.tsx
-packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/navbar.tsx
-packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/theme-switch.tsx
-packages/connect-explorer-theme/src/contexts/config.tsx > packages/connect-explorer-theme/src/constants.tsx > packages/connect-explorer-theme/src/components/toc.tsx
+packages/connect-explorer-theme/src/contexts/useConfig.tsx > packages/connect-explorer-theme/src/schema.ts > packages/connect-explorer-theme/src/components/navbar.tsx
+packages/connect-explorer-theme/src/contexts/useConfig.tsx > packages/connect-explorer-theme/src/schema.ts > packages/connect-explorer-theme/src/components/navbar.tsx > packages/connect-explorer-theme/src/utils/index.ts > packages/connect-explorer-theme/src/utils/use-git-edit-url.ts
+packages/connect-explorer-theme/src/contexts/useConfig.tsx > packages/connect-explorer-theme/src/schema.ts > packages/connect-explorer-theme/src/components/theme-switch.tsx
+packages/connect-explorer-theme/src/contexts/useConfig.tsx > packages/connect-explorer-theme/src/schema.ts > packages/connect-explorer-theme/src/components/toc.tsx
 packages/connect-explorer/src/actions/index.ts > packages/connect-explorer/src/actions/methodActions.ts > packages/connect-explorer/src/types/index.ts > packages/connect-explorer/src/actions/trezorConnectActions.ts
 packages/connect-explorer/src/actions/index.ts > packages/connect-explorer/src/actions/methodActions.ts > packages/connect-explorer/src/types/index.ts > packages/connect-explorer/src/reducers/index.ts > packages/connect-explorer/src/reducers/trezorConnectReducer.ts
 packages/connect-explorer/src/actions/methodActions.ts > packages/connect-explorer/src/types/index.ts
